+++ base/src/main/java/com/arnyminerz/escalaralcoiaicomtat/activity/MainActivity.kt
@@ -21,36 +21,36 @@ import com.arnyminerz.escalaralcoiaicomtat.activity.climb.AreaActivity
 import com.arnyminerz.escalaralcoiaicomtat.activity.isolated.EmailConfirmationActivity
 import com.arnyminerz.escalaralcoiaicomtat.activity.model.LanguageAppCompatActivity
 import com.arnyminerz.escalaralcoiaicomtat.activity.profile.ProfileActivity
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.AREAS
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.ENABLE_AUTHENTICATION
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.EXTRA_AREA
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.EXTRA_AREA_TRANSITION_NAME
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.LOCATION_PERMISSION_REQUEST_CODE
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.PREF_WAITING_EMAIL_CONFIRMATION
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.PREF_WARN_BATTERY
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.PROFILE_IMAGE_MAX_SIZE
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.SETTINGS_ALERT_PREF
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.TAB_ITEM_DOWNLOADS
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.TAB_ITEM_EXTRA
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.TAB_ITEM_HOME
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.TAB_ITEM_MAP
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.TAB_ITEM_SETTINGS
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.isLocationPermissionGranted
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.launch
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.putExtra
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.toast
+import com.arnyminerz.escalaralcoiaicomtat.core.view.getColorFromAttribute
+import com.arnyminerz.escalaralcoiaicomtat.core.view.visibility
 import com.arnyminerz.escalaralcoiaicomtat.databinding.ActivityMainBinding
 import com.arnyminerz.escalaralcoiaicomtat.fragment.DownloadsFragment
 import com.arnyminerz.escalaralcoiaicomtat.fragment.MapFragment
 import com.arnyminerz.escalaralcoiaicomtat.fragment.SettingsFragmentManager
 import com.arnyminerz.escalaralcoiaicomtat.fragment.climb.AreasViewFragment
 import com.arnyminerz.escalaralcoiaicomtat.fragment.preferences.MainSettingsFragment.Companion.SettingsPage
-import com.arnyminerz.escalaralcoiaicomtat.fragment.preferences.PREF_WAITING_EMAIL_CONFIRMATION
-import com.arnyminerz.escalaralcoiaicomtat.fragment.preferences.PREF_WARN_BATTERY
-import com.arnyminerz.escalaralcoiaicomtat.fragment.preferences.SETTINGS_ALERT_PREF
-import com.arnyminerz.escalaralcoiaicomtat.generic.isLocationPermissionGranted
-import com.arnyminerz.escalaralcoiaicomtat.generic.launch
-import com.arnyminerz.escalaralcoiaicomtat.generic.putExtra
-import com.arnyminerz.escalaralcoiaicomtat.generic.toast
 import com.arnyminerz.escalaralcoiaicomtat.intent.LoginRequestContract
 import com.arnyminerz.escalaralcoiaicomtat.list.adapter.MainPagerAdapter
-import com.arnyminerz.escalaralcoiaicomtat.shared.AREAS
-import com.arnyminerz.escalaralcoiaicomtat.shared.ENABLE_AUTHENTICATION
-import com.arnyminerz.escalaralcoiaicomtat.shared.EXTRA_AREA
-import com.arnyminerz.escalaralcoiaicomtat.shared.EXTRA_AREA_TRANSITION_NAME
-import com.arnyminerz.escalaralcoiaicomtat.shared.LOCATION_PERMISSION_REQUEST_CODE
-import com.arnyminerz.escalaralcoiaicomtat.shared.PROFILE_IMAGE_MAX_SIZE
 import com.arnyminerz.escalaralcoiaicomtat.shared.RESULT_CODE_LOGGED_IN
 import com.arnyminerz.escalaralcoiaicomtat.shared.RESULT_CODE_WAITING_EMAIL_CONFIRMATION
-import com.arnyminerz.escalaralcoiaicomtat.shared.TAB_ITEM_DOWNLOADS
-import com.arnyminerz.escalaralcoiaicomtat.shared.TAB_ITEM_EXTRA
-import com.arnyminerz.escalaralcoiaicomtat.shared.TAB_ITEM_HOME
-import com.arnyminerz.escalaralcoiaicomtat.shared.TAB_ITEM_MAP
-import com.arnyminerz.escalaralcoiaicomtat.shared.TAB_ITEM_SETTINGS
-import com.arnyminerz.escalaralcoiaicomtat.view.getColorFromAttribute
-import com.arnyminerz.escalaralcoiaicomtat.view.visibility
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.firebase.auth.FirebaseAuthInvalidUserException
 import com.google.firebase.auth.ktx.auth
