+++ base/src/main/java/com/arnyminerz/escalaralcoiaicomtat/activity/MapsActivity.kt
@@ -14,37 +14,37 @@ import androidx.activity.result.contract.ActivityResultContracts
 import androidx.annotation.UiThread
 import com.arnyminerz.escalaralcoiaicomtat.R
 import com.arnyminerz.escalaralcoiaicomtat.activity.model.LanguageAppCompatActivity
-import com.arnyminerz.escalaralcoiaicomtat.data.map.GeoGeometry
-import com.arnyminerz.escalaralcoiaicomtat.data.map.GeoMarker
-import com.arnyminerz.escalaralcoiaicomtat.data.map.MAP_LOAD_PADDING
-import com.arnyminerz.escalaralcoiaicomtat.data.map.getWindow
+import com.arnyminerz.escalaralcoiaicomtat.core.data.map.GeoGeometry
+import com.arnyminerz.escalaralcoiaicomtat.core.data.map.GeoMarker
+import com.arnyminerz.escalaralcoiaicomtat.core.data.map.MAP_LOAD_PADDING
+import com.arnyminerz.escalaralcoiaicomtat.core.data.map.getWindow
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.EXTRA_CENTER_CURRENT_LOCATION
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.EXTRA_KMZ_FILE
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.INFO_VIBRATION
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.LOCATION_PERMISSION_REQUEST_CODE
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.MAP_GEOMETRIES_BUNDLE_EXTRA
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.MAP_MARKERS_BUNDLE_EXTRA
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.MIME_TYPE_GPX
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.MIME_TYPE_KMZ
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.PERMISSION_DIALOG_TAG
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.SETTINGS_CENTER_MARKER_PREF
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.sharedPreferences
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.doAsync
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.fileName
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.getExtra
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.getParcelableList
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.isLocationPermissionGranted
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.maps.MapHelper
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.maps.MapNotInitializedException
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.maps.TrackMode
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.mime
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.toast
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.uiContext
 import com.arnyminerz.escalaralcoiaicomtat.databinding.ActivityMapsBinding
 import com.arnyminerz.escalaralcoiaicomtat.device.vibrate
 import com.arnyminerz.escalaralcoiaicomtat.fragment.dialog.BottomPermissionAskerFragment
-import com.arnyminerz.escalaralcoiaicomtat.fragment.preferences.SETTINGS_CENTER_MARKER_PREF
-import com.arnyminerz.escalaralcoiaicomtat.generic.doAsync
-import com.arnyminerz.escalaralcoiaicomtat.generic.fileName
-import com.arnyminerz.escalaralcoiaicomtat.generic.getExtra
-import com.arnyminerz.escalaralcoiaicomtat.generic.getParcelableList
-import com.arnyminerz.escalaralcoiaicomtat.generic.isLocationPermissionGranted
-import com.arnyminerz.escalaralcoiaicomtat.generic.maps.MapHelper
-import com.arnyminerz.escalaralcoiaicomtat.generic.maps.MapNotInitializedException
-import com.arnyminerz.escalaralcoiaicomtat.generic.maps.TrackMode
-import com.arnyminerz.escalaralcoiaicomtat.generic.mime
-import com.arnyminerz.escalaralcoiaicomtat.generic.toast
-import com.arnyminerz.escalaralcoiaicomtat.generic.uiContext
 import com.arnyminerz.escalaralcoiaicomtat.notification.DOWNLOAD_COMPLETE_CHANNEL_ID
 import com.arnyminerz.escalaralcoiaicomtat.notification.Notification
-import com.arnyminerz.escalaralcoiaicomtat.shared.EXTRA_CENTER_CURRENT_LOCATION
-import com.arnyminerz.escalaralcoiaicomtat.shared.EXTRA_KMZ_FILE
-import com.arnyminerz.escalaralcoiaicomtat.shared.INFO_VIBRATION
-import com.arnyminerz.escalaralcoiaicomtat.shared.LOCATION_PERMISSION_REQUEST_CODE
-import com.arnyminerz.escalaralcoiaicomtat.shared.MAP_GEOMETRIES_BUNDLE_EXTRA
-import com.arnyminerz.escalaralcoiaicomtat.shared.MAP_MARKERS_BUNDLE_EXTRA
-import com.arnyminerz.escalaralcoiaicomtat.shared.MIME_TYPE_GPX
-import com.arnyminerz.escalaralcoiaicomtat.shared.MIME_TYPE_KMZ
-import com.arnyminerz.escalaralcoiaicomtat.shared.PERMISSION_DIALOG_TAG
-import com.arnyminerz.escalaralcoiaicomtat.shared.sharedPreferences
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.ktx.firestore
