+++ base/src/main/java/com/arnyminerz/escalaralcoiaicomtat/activity/UpdatingActivity.kt
@@ -5,25 +5,25 @@ import androidx.annotation.WorkerThread
 import androidx.work.WorkInfo
 import com.arnyminerz.escalaralcoiaicomtat.R
 import com.arnyminerz.escalaralcoiaicomtat.activity.model.NetworkChangeListenerActivity
-import com.arnyminerz.escalaralcoiaicomtat.data.climb.area.Area
-import com.arnyminerz.escalaralcoiaicomtat.data.climb.dataclass.DataClass
-import com.arnyminerz.escalaralcoiaicomtat.data.climb.sector.Sector
-import com.arnyminerz.escalaralcoiaicomtat.data.climb.zone.Zone
+import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.area.Area
+import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.dataclass.DataClass
+import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.sector.Sector
+import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.zone.Zone
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.AREAS
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.ERROR_VIBRATE
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.QUIET_UPDATE
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.UPDATE_AREA
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.UPDATE_IMAGES
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.UPDATE_SECTOR
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.UPDATE_ZONE
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.doAsync
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.getExtra
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.launch
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.uiContext
+import com.arnyminerz.escalaralcoiaicomtat.core.view.visibility
 import com.arnyminerz.escalaralcoiaicomtat.databinding.ActivityUpdatingBinding
 import com.arnyminerz.escalaralcoiaicomtat.device.vibrate
-import com.arnyminerz.escalaralcoiaicomtat.generic.doAsync
-import com.arnyminerz.escalaralcoiaicomtat.generic.getExtra
-import com.arnyminerz.escalaralcoiaicomtat.generic.launch
-import com.arnyminerz.escalaralcoiaicomtat.generic.uiContext
 import com.arnyminerz.escalaralcoiaicomtat.network.base.ConnectivityProvider
-import com.arnyminerz.escalaralcoiaicomtat.shared.AREAS
-import com.arnyminerz.escalaralcoiaicomtat.shared.ERROR_VIBRATE
-import com.arnyminerz.escalaralcoiaicomtat.shared.QUIET_UPDATE
-import com.arnyminerz.escalaralcoiaicomtat.shared.UPDATE_AREA
-import com.arnyminerz.escalaralcoiaicomtat.shared.UPDATE_IMAGES
-import com.arnyminerz.escalaralcoiaicomtat.shared.UPDATE_SECTOR
-import com.arnyminerz.escalaralcoiaicomtat.shared.UPDATE_ZONE
-import com.arnyminerz.escalaralcoiaicomtat.view.visibility
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.ktx.firestore
