+++ base/src/main/java/com/arnyminerz/escalaralcoiaicomtat/activity/model/LanguageFragmentActivity.kt
@@ -2,7 +2,7 @@ package com.arnyminerz.escalaralcoiaicomtat.activity.model
 
 import android.content.Context
 import androidx.fragment.app.FragmentActivity
-import com.arnyminerz.escalaralcoiaicomtat.generic.loadLocale
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.loadLocale
 
 abstract class LanguageFragmentActivity : FragmentActivity() {
     @ExperimentalUnsignedTypes
