+++ base/src/main/java/com/arnyminerz/escalaralcoiaicomtat/auth/ProfileUtils.kt
@@ -8,13 +8,13 @@ import androidx.core.content.ContextCompat
 import androidx.core.graphics.drawable.toBitmap
 import androidx.core.graphics.scale
 import com.arnyminerz.escalaralcoiaicomtat.R
-import com.arnyminerz.escalaralcoiaicomtat.exception.CouldNotCompressImageException
-import com.arnyminerz.escalaralcoiaicomtat.generic.ValueMax
-import com.arnyminerz.escalaralcoiaicomtat.generic.WEBP_LOSSY_LEGACY
-import com.arnyminerz.escalaralcoiaicomtat.generic.cropToSquare
-import com.arnyminerz.escalaralcoiaicomtat.generic.doAsync
-import com.arnyminerz.escalaralcoiaicomtat.shared.PROFILE_IMAGE_COMPRESSION_QUALITY
-import com.arnyminerz.escalaralcoiaicomtat.shared.PROFILE_IMAGE_SIZE
+import com.arnyminerz.escalaralcoiaicomtat.core.exception.CouldNotCompressImageException
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.PROFILE_IMAGE_COMPRESSION_QUALITY
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.PROFILE_IMAGE_SIZE
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.ValueMax
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.WEBP_LOSSY_LEGACY
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.cropToSquare
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.doAsync
 import com.google.firebase.auth.FirebaseAuthInvalidUserException
 import com.google.firebase.auth.FirebaseUser
 import com.google.firebase.auth.ktx.userProfileChangeRequest
