+++ base/src/main/java/com/arnyminerz/escalaralcoiaicomtat/fragment/auth/RegisterFragment.kt
@@ -13,18 +13,18 @@ import com.arnyminerz.escalaralcoiaicomtat.activity.profile.AuthActivity
 import com.arnyminerz.escalaralcoiaicomtat.auth.createFirestoreUserReference
 import com.arnyminerz.escalaralcoiaicomtat.auth.setDefaultProfileImage
 import com.arnyminerz.escalaralcoiaicomtat.auth.updateDisplayName
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.CONFIRMATION_EMAIL_DYNAMIC
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.CONFIRMATION_EMAIL_URL
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.PREF_WAITING_EMAIL_CONFIRMATION
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.exception_handler.handleStorageException
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.doAsync
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.finishActivityWithResult
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.toast
+import com.arnyminerz.escalaralcoiaicomtat.core.utils.uiContext
+import com.arnyminerz.escalaralcoiaicomtat.core.view.viewListOf
+import com.arnyminerz.escalaralcoiaicomtat.core.view.visibility
 import com.arnyminerz.escalaralcoiaicomtat.databinding.FragmentAuthRegisterBinding
-import com.arnyminerz.escalaralcoiaicomtat.fragment.preferences.PREF_WAITING_EMAIL_CONFIRMATION
-import com.arnyminerz.escalaralcoiaicomtat.generic.doAsync
-import com.arnyminerz.escalaralcoiaicomtat.generic.finishActivityWithResult
-import com.arnyminerz.escalaralcoiaicomtat.generic.toast
-import com.arnyminerz.escalaralcoiaicomtat.generic.uiContext
-import com.arnyminerz.escalaralcoiaicomtat.list.viewListOf
-import com.arnyminerz.escalaralcoiaicomtat.shared.CONFIRMATION_EMAIL_DYNAMIC
-import com.arnyminerz.escalaralcoiaicomtat.shared.CONFIRMATION_EMAIL_URL
 import com.arnyminerz.escalaralcoiaicomtat.shared.RESULT_CODE_WAITING_EMAIL_CONFIRMATION
-import com.arnyminerz.escalaralcoiaicomtat.shared.exception_handler.handleStorageException
-import com.arnyminerz.escalaralcoiaicomtat.view.visibility
 import com.google.android.material.textfield.TextInputLayout
 import com.google.firebase.auth.ActionCodeSettings
 import com.google.firebase.auth.AuthResult
