+++ base/src/main/java/com/arnyminerz/escalaralcoiaicomtat/fragment/dialog/PathEquipmentDialog.kt
@@ -6,12 +6,12 @@ import android.os.Bundle
 import android.view.Window
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.arnyminerz.escalaralcoiaicomtat.R
-import com.arnyminerz.escalaralcoiaicomtat.data.climb.path.safes.FixedSafesData
-import com.arnyminerz.escalaralcoiaicomtat.data.climb.path.safes.RequiredSafesData
+import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.path.safes.FixedSafesData
+import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.path.safes.RequiredSafesData
+import com.arnyminerz.escalaralcoiaicomtat.core.view.visibility
 import com.arnyminerz.escalaralcoiaicomtat.databinding.DialogPathEquipmentBinding
 import com.arnyminerz.escalaralcoiaicomtat.list.adapter.EquipmentAdapter
 import com.arnyminerz.escalaralcoiaicomtat.list.adapter.EquipmentAdapterType
-import com.arnyminerz.escalaralcoiaicomtat.view.visibility
 
 class PathEquipmentDialog(
     context: Context,
