+++ base/src/main/java/com/arnyminerz/escalaralcoiaicomtat/fragment/preferences/NotificationsSettingsFragment.kt
@@ -9,6 +9,7 @@ import androidx.preference.Preference
 import androidx.preference.PreferenceFragmentCompat
 import androidx.preference.SwitchPreference
 import com.arnyminerz.escalaralcoiaicomtat.R
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.SETTINGS_ALERT_PREF
 
 class NotificationsSettingsFragment : PreferenceFragmentCompat() {
