+++ base/src/main/java/com/arnyminerz/escalaralcoiaicomtat/list/model/dwdataclass/DwDataClassViewHolder.kt
@@ -6,7 +6,7 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.arnyminerz.escalaralcoiaicomtat.R
-import com.arnyminerz.escalaralcoiaicomtat.data.climb.dataclass.DataClass
+import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.dataclass.DataClass
 import com.google.android.material.progressindicator.BaseProgressIndicator
 
 /**
