+++ base/src/main/java/com/arnyminerz/escalaralcoiaicomtat/shared/App.kt
@@ -5,6 +5,8 @@ import android.app.Application
 import android.content.Context
 import android.util.Log
 import com.arnyminerz.escalaralcoiaicomtat.BuildConfig
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.PREFERENCES_NAME
+import com.arnyminerz.escalaralcoiaicomtat.core.shared.sharedPreferences
 import com.arnyminerz.escalaralcoiaicomtat.network.base.ConnectivityProvider
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.ktx.auth
