+++ base/src/main/res/values-ca/strings.xml
@@ -263,10 +263,6 @@
     <string name="mark_completed_type_top_rope">Top-rope</string>
     <string name="mark_completed_type_required">El nombre d\'intents és necessari</string>
     <string name="mark_completed_type_invalid">Un tipus de marcatge vàlid és necessari</string>
-    <string name="notification_mark_completed_waiting_title">Programat marcatge com a completat</string>
-    <string name="notification_mark_completed_waiting_message">Esperant fins a tenir connexió a Internet per a enviar el comentari i marcar la via com a completada.</string>
-    <string name="notification_mark_completed_error_title">No s\'ha pogut marcar com a completada</string>
-    <string name="notification_mark_completed_error_message">La via \"%1$s\" ha sigut marcada com a completada, però ha ocorregut un erorr. Codi d\'error: %2$s</string>
     <string name="toast_error_account_manager_store">No s\'han pogut desar les dades del compte</string>
     <string name="action_google_auth">Accedir amb Google</string>
     <string name="toast_error_login_google">No s\'ha pogut accedir amb Google</string>
